// src/entities/tournaments.interface.ts
export enum Tournaments {
    UCL = 'UCL',
    EPL = 'EPL',
    LALIGA = 'LALIGA',
    SERIEA = 'SERIEA',
    BUNDESLIGA = 'BUNDESLIGA',
    LIGUE1 = 'LIGUE1',
    PRIMEIRALIGA = 'PRIMEIRALIGA',

    UKRAINELIGA = 'UKRAINELIGA',
    RUSSIANLIGA = 'RUSSIANLIGA',
    EREDIVISIE = 'EREDIVISIE',
    BELGIANLIGA = 'BELGIANLIGA',
    SCOTTISHLIGA = 'SCOTTISHLIGA',
    SAUDILIGA = 'SAUDILIGA',
    TURKISHLIGA = 'TURKISHLIGA',
}